AltaFactura Terminado
-Creacion y funcionamiento del form
-Modificacion de algunas entidades
-Conexion a la BD a travez de API

+++ TUP_P2_Hotel/HotelBackEnd/Model/FacturaDetalleModel.cs
@@ -12,7 +12,7 @@ namespace HotelBackEnd.Model
         public TipoServicioModel Servicio { get; set; }
         public int Cantidad { get; set; }
         public decimal Monto { get; set; }
-        public double CalcularSubTotal()
+        public decimal CalcularSubTotal()
         {
             return Monto * Cantidad;
         }
